+++ src/components/Home/SelectStartType.tsx
@@ -1,0 +1,20 @@
+const SelectStartType = () => {
+  return (
+    <div>
+      <a
+        href="#"
+        className="inline-flex items-center px-8 py-3 text-white transition bg-gray-900 rounded-full shadow-lg focus:outline-none focus:ring focus:bg-teal-600 hover:bg-gray-800"
+      >
+        <span className="text-sm font-medium"> 自己紹介作成 →</span>
+      </a>
+      <a
+        href="#"
+        className="ml-3 inline-flex items-center px-8 py-3 text-white transition bg-gray-900 rounded-full shadow-lg focus:outline-none focus:ring focus:bg-teal-600 hover:bg-gray-800"
+      >
+        <span className="text-sm font-medium"> カード取得 → </span>
+      </a>
+    </div>
+  );
+};
+
+export default SelectStartType;
