+++ src/hooks/useGeolocation.tsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useRef, useState } from "react";
+import { useCallback, useContext, useEffect, useRef, useState } from "react";
 import { GeolocationContext } from "../../pages/_app";
 
 export const useGeolocation = () => {
@@ -19,12 +19,12 @@ export const useGeolocation = () => {
     }
   }, [isAvailable]);
 
-  const getCurrentPosition = () => {
+  const getCurrentPosition = useCallback(() => {
     navigator.geolocation.getCurrentPosition((position) => {
       const { latitude, longitude } = position.coords;
       setPosition({ latitude: latitude, longitude: longitude });
     });
-  };
+  }, []);
 
   if (!isFirstRef && !isAvailable) {
     alert("位置情報が取得できないためアプリを続けることができません");
